Make TypeRegistry a single concrete class implementation
