+++ engine-tests/src/main/java/org/terasology/testUtil/ModuleManagerFactory.java
@@ -16,7 +16,6 @@
 package org.terasology.testUtil;
 
 import com.google.common.collect.Sets;
-import org.mockito.Mockito;
 import org.terasology.engine.TerasologyConstants;
 import org.terasology.engine.module.ModuleManager;
 import org.terasology.engine.module.ModuleManagerImpl;
@@ -24,11 +23,9 @@ import org.terasology.module.ClasspathModule;
 import org.terasology.module.ModuleMetadata;
 import org.terasology.module.ModuleMetadataReader;
 import org.terasology.naming.Name;
-import org.terasology.reflection.internal.TypeRegistryImpl;
 
 import java.io.InputStreamReader;
 import java.io.Reader;
-import java.util.Set;
 
 /**
  */
